backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Aleksey-Smirnov <smirnov@edsd.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
alembic = "^1.13.1"
fastapi = "^0.110.0"
pydantic = "^2.6.3"
pydantic-settings = "^2.2.1"
sqlalchemy = "^2.0.28"
psycopg2-binary = "^2.9.9"
databases = "^0.9.0"
asyncpg = "^0.29.0"
uvicorn = "^0.27.1"
pytest = "^8.0.2"
sqlalchemy-utils = "^0.41.1"
httpx = "^0.27.0"
pytest-asyncio = "^0.23.5.post1"
pytest-tornasync = "^0.6.0.post2"
pytest-trio = "^0.8.0"
pytest-twisted = "^1.14.0"
twisted = "^24.3.0"
minio = "^7.2.7"
python-multipart = "^0.0.9"
pycryptodome = "^3.20.0"
python-jose = "^3.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
